requirements.txt
apipkg==1.5
appnope==0.1.0
asgiref==3.2.7
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.4
decorator==4.4.2
Django==3.0.4
django-filter==2.2.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
drf-yasg==1.17.1
execnet==1.7.1
idna==2.9
importlib-metadata==1.5.2
inflection==0.3.1
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jedi==0.16.0
Jinja2==2.11.1
kafka-python==2.0.1
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.19.1
more-itertools==8.2.0
mysqlclient==1.4.6
packaging==20.3
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.4
protobuf==3.11.3
ptyprocess==0.6.0
py==1.8.1
Pygments==2.6.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pymongo==3.10.1
pyparsing==2.4.6
pytest==5.4.1
pytest-django==3.8.0
pytest-forked==1.1.3
pytest-xdist==1.31.0
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
uWSGI==2.0.18
wcwidth==0.1.9
wrapt==1.11.2
zipp==3.1.0